bot.py
import discord
from discord.ext import commands
import config
from keep_alive import keep_alive
from database import init_db  # ← Nuevo: inicializa SQLite
import asyncio

# Inicia el servidor Flask apenas arranca el bot
keep_alive()

# Lista de IDs de servidores a ignorar (cuando quieras, ponelos acá)
IGNORED_GUILDS = [
    # 123456789012345678,  # Ejemplo de servidor 1
    # 234567890123456789   # Ejemplo de servidor 2
]

intents = discord.Intents.default()
bot = commands.Bot(command_prefix="!", intents=intents)

@bot.event
async def on_ready():
    print(f"✅ Bot conectado como {bot.user}")

    # Establecer status personalizado
    activity = discord.Activity(type=discord.ActivityType.watching, name="2.941.013 players in Warzone")
    await bot.change_presence(status=discord.Status.online, activity=activity)

    # Mostrar lista de servidores en consola
    print("📜 Servidores donde está el bot:")
    for guild in bot.guilds:
        print(f"👉 {guild.name} (ID: {guild.id})")

    # Verificar y salirse de servidores no deseados
    for guild in bot.guilds:
        if guild.id in IGNORED_GUILDS:
            print(f"🚪 Saliendo del servidor: {guild.name}")
            await guild.leave()

    try:
        synced = await bot.tree.sync()
        print(f"🌐 Comandos sincronizados: {len(synced)}")
    except Exception as e:
        print(f"❌ Error al sincronizar comandos: {e}")

async def main():
    init_db()  # ← Este llamado crea la base si no existe
    await bot.load_extension("cogs.warzone")
    await bot.start(config.TOKEN)

asyncio.run(main())